requirements.txt
Flask~=2.3.2
tensorflow
numpy==1.24.3
pillow~=9.5.0
xgboost
scipy==1.10.1
scikit-learn==1.2.2
requests~=2.30.0
